+++ services/index.js
@@ -8,6 +8,7 @@ const RewardService = require('./reward-service');
 const StarService = require('./star-service');
 const TaskService = require('./task-service');
 const MessageService = require('./message-service');
+const OfflineQueueService = require('./offline-queue-service');
 const { UserService } = require('./user-service');
 const ValidationService = require('./validation-service');
 const ConfigService = require('./config-service');
@@ -24,6 +25,9 @@ module.exports = {
   
   // 消息服务
   MessageService,
+
+  // 离线队列服务
+  OfflineQueueService,
   
   // 用户服务
   UserService,